refactor: Refactored abstract CAM module
